+++ examples/with-rsc/src/components/Content.tsx
