+++ examples/with-rsc/src/components/EditButton.tsx
